+++ src/TenantiManager.php
@@ -33,14 +33,13 @@ class TenantiManager extends Manager
      */
     protected function createDriver($driver)
     {
-        $config = Arr::pull($this->config, "drivers.{$driver}");
-        $chunk  = Arr::pull($this->config, 'chunk', 100);
+        $chunk = Arr::get($this->config, 'chunk', 100);
 
-        if (is_null($config)) {
+        if (is_null($this->setupDriverConfig($driver))) {
             throw new InvalidArgumentException("Driver [$driver] not supported.");
         }
 
-        return $this->app->make($this->resolver, [$this->app, $driver, $config, $chunk]);
+        return $this->app->make($this->resolver, [$this->app, $this, $driver, $chunk]);
     }
 
     /**
@@ -56,11 +55,14 @@ class TenantiManager extends Manager
     /**
      * Get configuration values.
      *
-     * @return array
+     * @param  string|null  $group
+     * @param  mixed  $default
+     *
+     * @return mixed
      */
-    public function getConfig()
+    public function getConfig($group = null, $default = null)
     {
-        return $this->config;
+        return Arr::get($this->config, $group, $default);
     }
 
     /**
@@ -82,12 +84,13 @@ class TenantiManager extends Manager
      *
      * @param  string  $using
      * @param  \Closure  $callback
+     * @param  array  $option
      *
      * @return void
      *
      * @throws \InvalidArgumentException
      */
-    public function connection($using, Closure $callback)
+    public function connection($using, Closure $callback, array $options = [])
     {
         $repository = $this->app->make('config');
 
@@ -101,9 +104,11 @@ class TenantiManager extends Manager
             throw new InvalidArgumentException("Database connection [{$using}] is not available.");
         }
 
-        Arr::set($this->config, 'database', [
+        Arr::set($this->config, 'connection', [
+            'name'     => "{$using}_{id}",
             'template' => $config,
             'resolver' => $callback,
+            'options'  => $options,
         ]);
     }
 
@@ -122,5 +127,45 @@ class TenantiManager extends Manager
     public function setupMultiDatabase($using, Closure $callback)
     {
         return $this->connection($using, $callback);
+    }
+
+    /**
+     * Prepare configuration values.
+     *
+     * @param  string  $driver
+     *
+     * @return array|null
+     */
+    protected function setupDriverConfig($driver)
+    {
+        if (isset($this->config[$driver])) {
+            return;
+        }
+
+        if (is_null($config = Arr::pull($this->config, "drivers.{$driver}"))) {
+            return;
+        }
+
+        $connection = Arr::get($this->config, 'connection');
+
+        if (! is_null($connection) && $this->driverExcludedByOptions($driver, $connection['options'])) {
+            $connection = null;
+        }
+
+        return $this->config[$driver] = array_merge($config, ['connection' => $connection]);
+    }
+
+    /**
+     * Determine if the given options exclude a particular driver.
+     *
+     * @param  string  $driver
+     * @param  array  $options
+     *
+     * @return bool
+     */
+    protected function driverExcludedByOptions($driver, array $options)
+    {
+        return (! empty($options['only']) && ! in_array($driver, (array) $options['only'])) ||
+            (! empty($options['except']) && in_array($driver, (array) $options['except']));
     }
 }