+++ tests/Migrator/OperationTraitTest.php
@@ -2,6 +2,7 @@
 
 use Mockery as m;
 use Illuminate\Config\Repository;
+use Illuminate\Container\Container;
 use Illuminate\Database\Eloquent\Model;
 use Orchestra\Tenanti\Migrator\OperationTrait;
 
@@ -26,10 +27,11 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
     public function testAsDefaultDatabaseMethod()
     {
         $this->app = m::mock('\Illuminate\Container\Container[make]');
+        $this->driver = 'user';
 
         $repository = new Repository([
             'database' => [
-                'default' => 'mysql',
+                'default'     => 'mysql',
                 'connections' => [
                     'tenant' => [
                         'database' => 'tenants',
@@ -38,17 +40,20 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
             ],
         ]);
 
-        $this->config = [
-            'model' => 'User',
-            'database' => [
-                'template' => $repository->get('database.connections.tenant'),
-                'resolver' => function (Model $entity, array $template) {
-                    return array_merge($template, [
-                        'database' => "tenants_{$entity->getKey()}",
-                    ]);
-                },
-            ],
-        ];
+        $manager = m::mock('\Orchestra\Tenanti\TenantiManager', [$this->app]);
+
+        $manager->shouldReceive('getConfig')->with('user.connection', null)->andReturn([
+                    'template' => $repository->get('database.connections.tenant'),
+                    'resolver' => function (Model $entity, array $template) {
+                        return array_merge($template, [
+                            'database' => "tenants_{$entity->getKey()}",
+                        ]);
+                    },
+                    'name'    => 'tenant_{id}',
+                    'options' => ['only' => ['user']],
+                ]);
+
+        $this->manager = $manager;
 
         $model = m::mock('\Illuminate\Database\Eloquent\Model');
 
@@ -72,10 +77,11 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
     public function testResolveDatabaseConnectionMethod()
     {
         $this->app = m::mock('\Illuminate\Container\Container[make]');
+        $this->driver = 'user';
 
         $repository = new Repository([
             'database' => [
-                'default' => 'mysql',
+                'default'     => 'mysql',
                 'connections' => [
                     'tenant' => [
                         'database' => 'tenants',
@@ -84,17 +90,20 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
             ],
         ]);
 
-        $this->config = [
-            'model' => 'User',
-            'database' => [
+        $manager = m::mock('\Orchestra\Tenanti\TenantiManager', [$this->app]);
+
+        $manager->shouldReceive('getConfig')->with('user.connection', null)->andReturn([
                 'template' => $repository->get('database.connections.tenant'),
                 'resolver' => function (Model $entity, array $template) {
                     return array_merge($template, [
                         'database' => "tenants_{$entity->getKey()}",
                     ]);
                 },
-            ],
-        ];
+                'name'    => 'tenant_{id}',
+                'options' => [],
+            ]);
+
+        $this->manager = $manager;
 
         $model = m::mock('\Illuminate\Database\Eloquent\Model');
 
@@ -108,7 +117,6 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(['database' => 'tenants_5'], $repository->get('database.connections.tenant_5'));
     }
 
-
     /**
      * Test Orchestra\Tenanti\Migrator\OperationTrait::resolveModel()
      * method.
@@ -118,7 +126,14 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
     public function testResolveModelMethod()
     {
         $this->app = m::mock('\Illuminate\Container\Container[make]');
-        $this->config = ['model' => 'User'];
+        $this->driver = 'user';
+
+        $manager = m::mock('\Orchestra\Tenanti\TenantiManager', [$this->app]);
+
+        $manager->shouldReceive('getConfig')->with('user.model', null)->andReturn('User')
+            ->shouldReceive('getConfig')->with('user.database', null)->andReturnNull();
+
+        $this->manager = $manager;
 
         $model = m::mock('\Illuminate\Database\Eloquent\Model');
 
@@ -138,7 +153,14 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
     public function testResolveModelMethodWithConnectionName()
     {
         $this->app = m::mock('\Illuminate\Container\Container[make]');
-        $this->config = ['model' => 'User', 'database' => 'primary'];
+        $this->driver = 'user';
+
+        $manager = m::mock('\Orchestra\Tenanti\TenantiManager', [$this->app]);
+
+        $manager->shouldReceive('getConfig')->with('user.model', null)->andReturn('User')
+            ->shouldReceive('getConfig')->with('user.database', null)->andReturn('primary');
+
+        $this->manager = $manager;
 
         $model = m::mock('\Illuminate\Database\Eloquent\Model');
 
@@ -161,7 +183,13 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
     public function testResolveModelMethodThrowsException()
     {
         $this->app = m::mock('\Illuminate\Container\Container[make]');
-        $this->config = ['model' => 'User'];
+        $this->driver = 'user';
+
+        $manager = m::mock('\Orchestra\Tenanti\TenantiManager', [$this->app]);
+
+        $manager->shouldReceive('getConfig')->with('user.model', null)->andReturn('User');
+
+        $this->manager = $manager;
 
         $this->app->shouldReceive('make')->once()->with('User')->andReturnNull();
 
@@ -176,7 +204,14 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
      */
     public function testGetModelNameMethod()
     {
-        $this->config = ['model' => 'User'];
+        $app = new Container();
+        $this->driver = 'user';
+
+        $manager = m::mock('\Orchestra\Tenanti\TenantiManager', [$app]);
+
+        $manager->shouldReceive('getConfig')->with('user.model', null)->andReturn('User');
+
+        $this->manager = $manager;
 
         $this->assertEquals('User', $this->getModelName());
     }
@@ -189,8 +224,15 @@ class OperationTraitTest extends \PHPUnit_Framework_TestCase
      */
     public function testGetMigrationPathMethod()
     {
+        $this->driver = 'user';
         $path = realpath(__DIR__);
-        $this->config = ['path' => $path];
+        $app = new Container();
+
+        $manager = m::mock('\Orchestra\Tenanti\TenantiManager', [$app]);
+
+        $manager->shouldReceive('getConfig')->with('user.path', null)->andReturn($path);
+
+        $this->manager = $manager;
 
         $this->assertEquals($path, $this->getMigrationPath());
     }